+++ src/housekeeping.hh
@@ -1,0 +1,17 @@
+#include "pinocchio_cppadcg.hh"
+#include "lang_cpp.hh"
+#include "lang_rust.hh"
+
+
+using namespace pinocchio;
+using namespace CppAD;
+using namespace CppAD::cg;
+
+// Typedef for AD types
+using CGD = CG<double>;
+using ADCG = AD<CGD>;
+
+using ADModel = ModelTpl<ADCG>;
+using ADData = DataTpl<ADCG>;
+using ADVectorXs = Eigen::Matrix<ADCG, Eigen::Dynamic, 1>;
+using ADMatrixXs = Eigen::Matrix<ADCG, Eigen::Dynamic, Eigen::Dynamic>;
